Committed a few files

+++ core/src/com/hycap/othello/AdvantageOnlyEvaluator.java
@@ -2,7 +2,7 @@ package com.hycap.othello;
 
 public class AdvantageOnlyEvaluator implements Evaluator{
     @Override
-    public int Evaluate(Board board) {
+    public float Evaluate(Board board) {
         return board.getWhiteAdvantage();
     }
 }
