+++ core/src/com/hycap/othello/Evaluator.java
@@ -2,5 +2,5 @@ package com.hycap.othello;
 
 // Always returns white's advantage
 public interface Evaluator {
-    int Evaluate(Board board);
+    float Evaluate(Board board);
 }
